requirements.txt
faiss-cpu==1.7.4
fastapi==0.105.0
ffmpeg-python==0.2.0
langchain==0.0.352
openai==1.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
soundfile==0.12.1
tiktoken==0.5.2
uvicorn==0.25.0
websockets==12.0
